feat: implement file upload with user id
